Decrease duplication and support switching redisdb with shell scripts
